Snapshot local working changes

+++ lib/matplotlib/hatch.py
@@ -227,13 +227,13 @@ def get_path(hatchpattern, density=6):
     if num_vertices == 0:
         return Path(mx.zeros((0, 2)))
 
-    vertices = mx.zeros((num_vertices, 2))
+    vertices = mx.zeros((num_vertices, 2), dtype=mx.float64)
     codes = mx.zeros(num_vertices, Path.code_type)
 
     cursor = 0
     for pattern in patterns:
         if pattern.num_vertices != 0:
-            vertices_chunk = mx.zeros((pattern.num_vertices, 2))
+            vertices_chunk = mx.zeros((pattern.num_vertices, 2), dtype=mx.float64)
             codes_chunk = mx.zeros(pattern.num_vertices, Path.code_type)
             pattern.set_vertices_and_codes(vertices_chunk, codes_chunk)
             vertices[cursor:cursor + pattern.num_vertices] = vertices_chunk
